feat(Overview): refactor summary, move to powerpack
