--- shared/src/containers/ProjectTreeTable/components/TableFooter/index.ts
@@ -1,8 +1,0 @@
-export { TableFooter } from './TableFooter'
-export { classifyColumnSummary } from './classifyColumnSummary'
-export { useColumnSummaries } from './useColumnSummaries'
-export type { ColumnSummariesResult } from './useColumnSummaries'
-export { mapColumnStatsToSummary } from './mapColumnStats'
-export type { FieldStats } from './mapColumnStats'
-export { mockFieldStats } from './mockFieldStats'
-export * from './summaryTypes'
