+++ shared/src/containers/ProjectTreeTable/components/index.ts
@@ -4,4 +4,3 @@ export { default as RowSelectionHeader } from './RowSelectionHeader'
 export { SelectionCell } from './SelectionCell'
 export { ColumnHeaderMenu } from './ColumnHeaderMenu'
 export { default as ColumnDndProvider } from './ColumnDndProvider'
-export { mockFieldStats } from './TableFooter'
